third_party/blink/web_tests/platform/mac/external/wpt/uievents/mouse/mouse_boundary_events_after_reappending_last_over_target.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] After re-appending the last over element to outside the deepest mouseenter target (but keeps it as under the cursor) at mouseover, mouse boundary events should be fired on it again to correct the following mouseleave targets
  assert_equals: expected "mouseover@div#child, mouseenter@body, mouseenter@div#grandparent, mouseenter@div#parent, mouseenter@div#child, mousemove@div#child, mouseleave@div#parent, mouseover@div#child, mouseenter@div#child" but got "mouseover@div#child, mouseenter@body, mouseenter@div#grandparent, mouseenter@div#parent, mouseenter@div#child, mousemove@div#child"
[FAIL] After re-appending the last over element to outside the deepest mouseenter target at mouseover, mouse boundary events should be fired only on the element under the cursor
  assert_equals: expected "mouseover@div#child, mouseenter@body, mouseenter@div#grandparent, mouseenter@div#parent, mouseenter@div#child, mousemove@div#child, mouseover@div#parent" but got "mouseover@div#child, mouseenter@body, mouseenter@div#grandparent, mouseenter@div#parent, mouseenter@div#child, mousemove@div#child, mouseout@div#child, mouseleave@div#child, mouseover@div#parent, mouseenter@div#parent"
[FAIL] After re-appending the last over element to outside the deepest mouseenter target (but keeps under the cursor) after mouseover, mouse boundary events should be fired on it again to correct the following mouseleave event targets
  assert_equals: expected "mouseover@div#child, mouseenter@body, mouseenter@div#grandparent, mouseenter@div#parent, mouseenter@div#child, mousemove@div#child, mouseleave@div#parent, mouseover@div#child, mouseenter@div#child" but got "mouseover@div#child, mouseenter@body, mouseenter@div#grandparent, mouseenter@div#parent, mouseenter@div#child, mousemove@div#child"
[FAIL] After re-appending the last over element to outside the deepest mouseenter target after mouseover, mouse boundary events should be fired only on the element under the cursor
  assert_equals: expected "mouseover@div#child, mouseenter@body, mouseenter@div#grandparent, mouseenter@div#parent, mouseenter@div#child, mousemove@div#child, mouseover@div#parent" but got "mouseover@div#child, mouseenter@body, mouseenter@div#grandparent, mouseenter@div#parent, mouseenter@div#child, mousemove@div#child, mouseout@div#child, mouseleave@div#child, mouseover@div#parent, mouseenter@div#parent"
[FAIL] After removing after mouseover and re-appending the last over element at next animation frame, mouse boundary events should be fired
  assert_equals: expected "mouseover@div#child, mouseenter@body, mouseenter@div#grandparent, mouseenter@div#parent, mouseenter@div#child, mousemove@div#child, mouseover@div#parent, mouseout@div#parent, mouseover@div#child, mouseenter@div#child" but got "mouseover@div#child, mouseenter@body, mouseenter@div#grandparent, mouseenter@div#parent, mouseenter@div#child, mousemove@div#child, mouseover@div#parent, mouseenter@body, mouseenter@div#grandparent, mouseenter@div#parent, mouseout@div#parent, mouseover@div#child, mouseenter@div#child"
Harness: the test ran to completion.
